Updated the XBox port Added some pretty icons to PSP Added loads of runtime PSP diagnistic stuff Fixed GCC compiler warnings
git-svn-id: https://fuji.googlecode.com/svn/trunk@188 7e75b669-355e-e21c-2746-f45317fb21cf

+++ Fuji/Private/Source/Materials/Mat_Standard.cpp
@@ -241,7 +241,7 @@ int Mat_Standard_GetNumParams()
 
 MFMaterialParamaterInfo* Mat_Standard_GetParamaterInfo(int paramaterIndex)
 {
-	DBGASSERT(paramaterIndex < sizeof(paramaterInformation)/sizeof(MFMaterialParamaterInfo), STR("Invalid paramater id %d.", paramaterIndex));
+	DBGASSERT((uint32)paramaterIndex < sizeof(paramaterInformation)/sizeof(MFMaterialParamaterInfo), STR("Invalid paramater id %d.", paramaterIndex));
 
 	return &paramaterInformation[paramaterIndex];
 }
